solve bug int to float for rating review

+++ public/js/app.js
@@ -2834,7 +2834,7 @@ __webpack_require__.r(__webpack_exports__);
 
       if (ListOrg.review.length != 0) {
         for (var i = 0; i < ListOrg.review.length; i++) {
-          rating += parseInt(ListOrg.review[i].rating);
+          rating += parseFloat(ListOrg.review[i].rating);
         }
 
         avg = rating / ListOrg.review.length;
@@ -53710,15 +53710,14 @@ __webpack_require__.r(__webpack_exports__);
 /*!**********************************************************!*\
   !*** ./resources/js/components/ListOrg/ShowOrgGuest.vue ***!
   \**********************************************************/
-/*! no static exports found */
+/*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _ShowOrgGuest_vue_vue_type_template_id_061134f6___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./ShowOrgGuest.vue?vue&type=template&id=061134f6& */ "./resources/js/components/ListOrg/ShowOrgGuest.vue?vue&type=template&id=061134f6&");
 /* harmony import */ var _ShowOrgGuest_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./ShowOrgGuest.vue?vue&type=script&lang=js& */ "./resources/js/components/ListOrg/ShowOrgGuest.vue?vue&type=script&lang=js&");
-/* harmony reexport (unknown) */ for(var __WEBPACK_IMPORT_KEY__ in _ShowOrgGuest_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__) if(__WEBPACK_IMPORT_KEY__ !== 'default') (function(key) { __webpack_require__.d(__webpack_exports__, key, function() { return _ShowOrgGuest_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__[key]; }) }(__WEBPACK_IMPORT_KEY__));
-/* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
+/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
 
 
 
@@ -53748,7 +53747,7 @@ component.options.__file = "resources/js/components/ListOrg/ShowOrgGuest.vue"
 /*!***********************************************************************************!*\
   !*** ./resources/js/components/ListOrg/ShowOrgGuest.vue?vue&type=script&lang=js& ***!
   \***********************************************************************************/
-/*! no static exports found */
+/*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
